refactor: Make base and provider classes separated
