+++ source/internal/provider.ts
@@ -1,21 +1,10 @@
-import { RESOURCE_DOMAIN } from './constants';
-import type { Hitomi } from '../hitomi';
-import type { Node } from './types';
-import { compare, defineProperties } from './functions';
-import { HitomiError } from '../structures/error';
 import { ResponseType } from '@platform';
-
-export class Base {
-	// @internal
-	protected readonly hitomi!: Hitomi;
-
-	// @internal
-	constructor(hitomi: Hitomi) {
-		defineProperties(this, {
-			hitomi: hitomi
-		});
-	}
-}
+import type { Hitomi } from '../hitomi';
+import { HitomiError } from '../structures/error';
+import { Base } from './base';
+import { RESOURCE_DOMAIN } from './constants';
+import { compare } from './functions';
+import type { Node } from './types';
 
 // @internal
 export class Provider<T> extends Base {
@@ -32,8 +21,8 @@ export class Provider<T> extends Base {
 	}
 
 	public async retrieve(): Promise<T> {
-		if(Date.now() > this['expiresAt']) {
-			if(this['promise']) {
+		if (Date.now() > this['expiresAt']) {
+			if (this['promise']) {
 				return this['promise'];
 			}
 
@@ -63,7 +52,7 @@ export class IndexProvider extends Provider<string> {
 	public async getNodeAtAddress(address: Node[2][number], version: string): Promise<Node | undefined> {
 		const view: DataView = await this['hitomi'].request(RESOURCE_DOMAIN, '/' + this['field'] + 'index/' + this['field'] + '.' + version + '.index', ResponseType['VIEW'], address + '-' + (address + 463n));
 
-		if(!view['byteLength']) {
+		if (!view['byteLength']) {
 			return;
 		}
 
@@ -73,10 +62,10 @@ export class IndexProvider extends Provider<string> {
 		let offset: number = 4;
 		let i: number;
 
-		for(i = 0; i < keyCount; i++) {
+		for (i = 0; i < keyCount; i++) {
 			const keySize: number = view.getInt32(offset);
 
-			if(keySize < 1 || keySize > 31) {
+			if (keySize < 1 || keySize > 31) {
 				throw new HitomiError('KeySize', 'between 1 and 31');
 			}
 
@@ -89,13 +78,13 @@ export class IndexProvider extends Provider<string> {
 
 		offset += 4;
 
-		for(i = 0; i < dataCount; i++) {
+		for (i = 0; i < dataCount; i++) {
 			node[1].push([view.getBigUint64(offset), view.getInt32(offset + 8)]);
 
 			offset += 12;
 		}
 
-		for(i = 0; i < 17; i++) {
+		for (i = 0; i < 17; i++) {
 			node[2].push(view.getBigUint64(offset));
 
 			offset += 8;
@@ -105,45 +94,43 @@ export class IndexProvider extends Provider<string> {
 	}
 
 	public async binarySearch(key: Uint8Array, node: Node, version: string): Promise<Node[1][number] | undefined> {
-		if(!node[0]['length']) {
+		if (!node[0]['length']) {
 			return;
 		}
 
 		let compareResult: number = -1;
 		let index: number = 0;
 
-		while(
-			index < node[0]['length'] &&
-			(compareResult = compare(key, node[0][index])) === 1
-		) {
+		while (index < node[0]['length'] &&
+			(compareResult = compare(key, node[0][index])) === 1) {
 			index++;
 		}
 
-		if(!compareResult) {
+		if (!compareResult) {
 			return node[1][index];
 		}
 
-		if(!node[2][index]) {
+		if (!node[2][index]) {
 			return;
 		}
 
 		let isLeaf: boolean = true;
 
-		for(let i: number = 0; i < node[2]['length']; i++) {
-			if(node[2][i]) {
+		for (let i: number = 0; i < node[2]['length']; i++) {
+			if (node[2][i]) {
 				isLeaf = false;
 
 				break;
 			}
 		}
 
-		if(isLeaf) {
+		if (isLeaf) {
 			return;
 		}
 
 		const nextNode: Node | undefined = await this.getNodeAtAddress(node[2][index], version);
 
-		if(!nextNode) {
+		if (!nextNode) {
 			return;
 		}
