+++ source/managers/tag.ts
@@ -2,7 +2,7 @@ import { ResponseType } from '@platform';
 import { HitomiError } from '../structures/error';
 import type { Hitomi } from '../hitomi';
 import { TAG_TYPES, TAG_INDEX_DOMAIN, GALLERY_TYPES, LANGUAGE_NAMES, FRONT_DOMAIN } from '../internal/constants';
-import { Base } from '../internal/structures';
+import { Base } from '../internal/base';
 import { Tag } from '../structures/tag';
 
 /**
