+++ source/structures/media.ts
@@ -1,7 +1,7 @@
 import type { Hitomi } from '../hitomi';
 import { BASE_DOMAIN } from '../internal/constants';
 import { capitalize } from '../internal/functions';
-import { Base } from '../internal/structures';
+import { Base } from '../internal/base';
 import { HitomiError } from './error';
 import type { ImageContext } from '../internal/types';
 import type { Gallery } from './gallery';
