+++ tests/shared/constants.ts
@@ -1,4 +1,4 @@
-import { TAG_TYPES } from '../../source/internal/constants';
+import { TAG_TYPES } from '@/internal/constants';
 
 export const PARTIAL_TAG_TYPES: typeof TAG_TYPES = new Set(TAG_TYPES);
